Find the current user's GitHub username

Uses the GITHUB_USERNAME global variable or searches on GitHub, for the user's email address, see email_address.

Usage

gh_username(token = NULL, fallback = NULL)

Arguments

token

GitHub token to use. By default it uses the GITHUB_TOKEN environment variable, if set. If unset, uses the GITHUB_PAT environment variable, if set.

fallback

If not NULL then this value is returned if the GitHub username cannot be found, instead of triggering an error.

Value

GitHub username, or an error is thrown if it cannot be found.

Details

This function caches the username in the current R session, and if the email address of the user is unchanged, it does not perform another GitHub query.

See also

Examples

if (FALSE) { # \dontrun{
gh_username()
} # }